add a footer
