+++ pkg/api/helpers.go
@@ -4,10 +4,8 @@ import (
 	"compress/gzip"
 	"encoding/gob"
 	"encoding/json"
-	"html/template"
 	"io"
 	"net/http"
-	"path/filepath"
 	"strings"
 )
 
@@ -33,23 +31,6 @@ func readBody(dest io.Writer, r *http.Request) bool {
 			logger.WithError(err).Error("r.body.Read() failed")
 			return false
 		}
-	}
-	return true
-}
-
-func parseAndExecute(dest io.Writer, data interface{}, templateFiles ...string) bool {
-	templateFiles = append(templateFiles,
-		filepath.Join(PublicFiles, "header.gohtml"),
-		filepath.Join(PublicFiles, "navbar.gohtml"),
-	)
-	uploadTemplate, err := template.ParseFiles(templateFiles...)
-	if err != nil {
-		logger.WithError(err).Error("template.ParseFiles() failed")
-		return false
-	}
-	if err := uploadTemplate.Execute(dest, &data); err != nil {
-		logger.WithError(err).Error("template.Execute() failed")
-		return false
 	}
 	return true
 }
