+++ public/backups.gohtml
@@ -62,5 +62,6 @@
         </div>
     </div>
 </main>
+{{ template "footer.gohtml" }}
 </body>
 </html>
