+++ public/footer.gohtml
@@ -1,0 +1,14 @@
+<footer class="footer">
+    <div class="container mt-3 text-center">
+        <a class="text-light" href="https://www.facebook.com/groups/1080154885682377/">
+            <i class="fab fa-facebook-square fa-2x"></i>
+        </a>
+        <a class="text-light" href="https://github.com/virtual-vgo/vvgo">
+            <i class="fab fa-github-square fa-2x"></i>
+        </a>
+        <a class="text-light"
+           href="https://www.youtube.com/channel/UCeipEtsfjAA_8ATsd7SNAaQ">
+            <i class="fab fa-youtube-square fa-2x"></i>
+        </a>
+    </div>
+</footer>
