+++ public/index.gohtml
@@ -36,5 +36,6 @@
         </div>
     </div>
 </div>
+{{ template "footer.gohtml" }}
 </body>
 </html>
