+++ public/parts.gohtml
@@ -86,5 +86,6 @@
         </div>
     </div>
 </main>
+{{ template "footer.gohtml" }}
 </body>
 </html>
